Inline SymCrypt SPKI key import
